+++ xds/src/test/java/io/grpc/xds/orca/OrcaServiceImplTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.grpc.xds;
+package io.grpc.xds.orca;
 
 import static com.google.common.truth.Truth.assertThat;
 import static org.mockito.ArgumentMatchers.any;
@@ -40,6 +40,7 @@ import io.grpc.inprocess.InProcessChannelBuilder;
 import io.grpc.inprocess.InProcessServerBuilder;
 import io.grpc.internal.FakeClock;
 import io.grpc.testing.GrpcCleanupRule;
+import io.grpc.xds.orca.OrcaOobService;
 import java.util.Iterator;
 import java.util.Random;
 import java.util.concurrent.CyclicBarrier;
